+++ platform/ios/src/MGLAnnotationView.m
@@ -27,6 +27,19 @@
     // Intentionally left blank. The default implementation of this method does nothing.
 }
 
+- (void)setCenterOffset:(CGVector)centerOffset
+{
+    _centerOffset = centerOffset;
+    self.center = self.center;
+}
+
+- (void)setCenter:(CGPoint)center
+{
+    center.x += _centerOffset.dx;
+    center.y += _centerOffset.dy;
+    [super setCenter:center];
+}
+
 - (id<CAAction>)actionForLayer:(CALayer *)layer forKey:(NSString *)event
 {
     // Allow mbgl to drive animation of this view’s bounds.